+++ packages/ubcli/lib/ddlGenerators/db/MSSQL2012.js
@@ -242,7 +242,7 @@ ORDER BY i.object_id, c.name`
   /** @override */
   genCodeSetCaption (tableName, column, value, oldValue) {
     if (value) value = value.replace(/'/g, "''")
-    const proc = oldValue === null  ? 'sp_addextendedproperty': 'sp_updateextendedproperty'
+    const proc = oldValue === null ? 'sp_addextendedproperty' : 'sp_updateextendedproperty'
     let result = `EXEC ${proc} @name = N'${DB_DESCRIPTION_PROPERTY}', @value = N'${value === null ? (column || tableName) : value}',@level0type = N'SCHEMA',  @level0name= N'dbo', @level1type = N'TABLE',  @level1name = N'${tableName}'`
     if (column) result += `, @level2type = N'Column', @level2name = '${column}'`
     this.DDL.caption.statements.push(result)
@@ -374,6 +374,16 @@ ORDER BY i.object_id, c.name`
   }
 
   /** @override */
+  genCodeEnableMultitenancy (table) {
+    throw new Error('multitenancy is not implemented for MSSQL')
+  }
+
+  /** @override */
+  genCodeDisableMultitenancy (table) {
+    throw new Error('multitenancy is not implemented for MSSQL')
+  }
+
+  /** @override */
   genCodeCreatePK (table) {
     this.DDL.createPK.statements.push(
       `alter table dbo.${table.name} add constraint ${table.primaryKey.name} PRIMARY KEY CLUSTERED(${table.primaryKey.keys.join(',')})`
@@ -385,7 +395,7 @@ ORDER BY i.object_id, c.name`
     if (!constraintFK.generateFK) return
 
     const refTo = _.find(this.refTableDefs, { _nameUpper: constraintFK.references.toUpperCase() })
-    const refKeys = refTo ? refTo.primaryKey.keys.join(',') : 'ID'
+    const refKeys = refTo ? refTo.primaryKey.keys.join(',') : constraintFK.refPkDefColumn
 
     this.DDL.createFK.statements.push(
       `alter table dbo.${table.name} add constraint ${constraintFK.name} foreign key (${constraintFK.keys.join(',')}) references dbo.${constraintFK.references}(${refKeys})`